chore: Moved Dag Renderer to its own class.
